Fixed tests for latest changes; restructured test files
