+++ UnderDevelopment/tests/test_transformer_regulator.py
@@ -1,15 +1,15 @@
 from GridCal.Engine.PowerFlowDriver import PowerFlowOptions, SolverType, ReactivePowerControlMode
-from GridCal.Engine.PowerFlowDriver import PowerFlow, TapsControlMode
+from GridCal.Engine.PowerFlowDriver import PowerFlowMP, TapsControlMode
 from GridCal.Engine.CalculationEngine import MultiCircuit
 from GridCal.Engine.Devices import *
 from GridCal.Engine.DeviceTypes import *
 
-test_name = "test_pv_3"
 Sbase = 100 # MVA
 
-def complexe(z, XR):
+
+def complex_impedance(z, XR):
     """
-    Returns the complex impedance from z (in %) and the X/R ratio.
+    Returns the complex impedance from z and the X/R ratio.
     """
     z = float(abs(z))
     XR = float(abs(XR))
@@ -20,19 +20,12 @@ def complexe(z, XR):
         imag = 0.0
     return complex(real, imag)
 
-def test_pv_3():
-    """
-    Voltage controlled generator test, also useful for a basic tutorial. In this
-    case the generator M32 regulates the voltage at a setpoint of 1.025 pu, and
-    the slack bus (POI) regulates it at 1.0 pu.
 
-    The transformers' magnetizing branch losses are considered, as well as the
-    main power transformer's voltage regulator (X_C3) which regulates bus
-    B_MV_M32 at 1.005 pu.
-
-    In addition, the iterative PV control method is used instead of the usual
-    (faster) method.
+def test_gridcal_regulator():
     """
+    GridCal test for the new implementation of transformer voltage regulators.
+    """
+    test_name = "test_gridcal_regulator"
     grid = MultiCircuit(name=test_name)
     grid.Sbase = Sbase
     grid.time_profile = None
@@ -43,6 +36,10 @@ def test_pv_3():
               vnom=100, #kV
               is_slack=True)
     grid.add_bus(POI)
+
+    B_C3 = Bus(name="B_C3",
+               vnom=10) #kV
+    grid.add_bus(B_C3)
 
     B_MV_M32 = Bus(name="B_MV_M32",
                    vnom=10) #kV
@@ -57,13 +54,11 @@ def test_pv_3():
     UT.bus = POI
     grid.add_generator(POI, UT)
 
-    M32 = Generator(name="M32",
-                    active_power=4.2,
-                    voltage_module=1.025,
-                    Qmin=-2.5,
-                    Qmax=2.5)
+    # Create static generators (with fixed power factor)
+    M32 = StaticGenerator(name="M32",
+                          P=4.2, Q=0.0)  # MVA (complex)
     M32.bus = B_LV_M32
-    grid.add_generator(B_LV_M32, M32)
+    grid.add_static_generator(B_LV_M32, M32)
 
     # Create transformer types
     s = 100 # MVA
@@ -72,11 +67,11 @@ def test_pv_3():
     SS = TransformerType(name="SS",
                          hv_nominal_voltage=100, # kV
                          lv_nominal_voltage=10, # kV
-                         nominal_power=s,
-                         copper_losses=complexe(z, xr).real*s*1000/Sbase,
+                         nominal_power=s, # MVA
+                         copper_losses=complex_impedance(z, xr).real*s*1000/Sbase, # kW
                          iron_losses=125, # kW
                          no_load_current=0.5, # %
-                         short_circuit_voltage=z)
+                         short_circuit_voltage=z) # %
     grid.add_transformer_type(SS)
 
     s = 5 # MVA
@@ -85,24 +80,31 @@ def test_pv_3():
     PM = TransformerType(name="PM",
                          hv_nominal_voltage=10, # kV
                          lv_nominal_voltage=0.6, # kV
-                         nominal_power=s,
-                         copper_losses=complexe(z, xr).real*s*1000/Sbase,
+                         nominal_power=s, # MVA
+                         copper_losses=complex_impedance(z, xr).real*s*1000/Sbase, # kW
                          iron_losses=6.25, # kW
                          no_load_current=0.5, # %
-                         short_circuit_voltage=z)
+                         short_circuit_voltage=z) # %
     grid.add_transformer_type(PM)
 
     # Create branches
     X_C3 = Branch(bus_from=POI,
-                  bus_to=B_MV_M32,
+                  bus_to=B_C3,
                   name="X_C3",
                   branch_type=BranchType.Transformer,
                   template=SS,
                   bus_to_regulated=True,
-                  vset=1.005)
+                  vset=1.05)
     X_C3.tap_changer = TapChanger(taps_up=16, taps_down=16, max_reg=1.1, min_reg=0.9)
     X_C3.tap_changer.set_tap(X_C3.tap_module)
     grid.add_branch(X_C3)
+
+    C_M32 = Branch(bus_from=B_C3,
+                   bus_to=B_MV_M32,
+                   name="C_M32",
+                   r=7.84,
+                   x=1.74)
+    grid.add_branch(C_M32)
 
     X_M32 = Branch(bus_from=B_MV_M32,
                    bus_to=B_LV_M32,
@@ -124,16 +126,16 @@ def test_pv_3():
                                verbose=True,
                                initialize_with_existing_solution=True,
                                multi_core=True,
-                               control_q=ReactivePowerControlMode.Iterative,
+                               control_q=ReactivePowerControlMode.Direct,
                                control_taps=TapsControlMode.Direct,
                                tolerance=1e-6,
                                max_iter=99)
 
-    power_flow = PowerFlow(grid, options)
+    power_flow = PowerFlowMP(grid, options)
     power_flow.run()
 
     approx_volt = [round(100*abs(v), 1) for v in power_flow.results.voltage]
-    solution = [100.0, 100.7, 102.5] # Expected solution from GridCal
+    solution = [100.0, 105.2, 130.0, 130.1] # Expected solution from GridCal
 
     print()
     print(f"Test: {test_name}")
@@ -143,17 +145,12 @@ def test_pv_3():
 
     print("Generators:")
     for g in grid.get_generators():
-        print(f" - Generator {g}: q_min={g.Qmin} MVAR, q_max={g.Qmax} MVAR")
+        print(f" - Generator {g}: q_min={g.Qmin}pu, q_max={g.Qmax}pu")
     print()
 
     print("Branches:")
     for b in grid.branches:
-        print(f" - {b}:")
-        print(f"   R = {round(b.R, 4)} pu")
-        print(f"   X = {round(b.X, 4)} pu")
-        print(f"   X/R = {round(b.X/b.R, 1)}")
-        print(f"   G = {round(b.G, 4)} pu")
-        print(f"   B = {round(b.B, 4)} pu")
+        print(f" - {b}: R={round(b.R, 4)}pu, X={round(b.X, 4)}pu, X/R={round(b.X/b.R, 1)}, vset={b.vset}")
     print()
 
     print("Transformer types:")
@@ -163,8 +160,14 @@ def test_pv_3():
 
     print("Losses:")
     for i in range(len(grid.branches)):
-        print(f" - {grid.branches[i]}: losses={1000*round(power_flow.results.losses[i], 3)} kVA")
+        print(f" - {grid.branches[i]}: losses={round(power_flow.results.losses[i], 3)} MVA")
     print()
+
+    print(f"Voltage settings: {grid.numerical_circuit.vset}")
+
+    #print("GridCal logger:")
+    #for l in grid.logger:
+        #print(f" - {l}")
 
     equal = True
     for i in range(len(approx_volt)):
